+++ docsig/_main.py
@@ -13,8 +13,6 @@ from ._core import get_files as _get_files
 from ._core import get_members as _get_members
 from ._core import populate as _populate
 from ._core import print_failures as _print_failures
-from ._function import Function as _Function
-from ._report import warn as _warn
 
 
 def main() -> int:
@@ -24,20 +22,18 @@ def main() -> int:
     """
     paths: _t.List[_Path] = []
     failures: _FailedDocData = {}
-    missing: _t.List[_t.Tuple[str, _Function]] = []
     parser = _Parser()
     for path in parser.args.path:
         _get_files(path, paths)
 
     members = _get_members(paths)
     for module in members:
-        _populate(module.name, module, failures, missing)
+        _populate(module.name, module, failures)
         for klass in module.classes:
             name = f"{module.name}::{klass.name}"
-            _populate(name, klass, failures, missing)
+            _populate(name, klass, failures)
 
     _print_failures(failures)
-    _warn(missing)
 
     # pylint: disable=use-implicit-booleaness-not-comparison
     return int(failures != {})